+++ tasks/wrappers/browserify.js
@@ -1,10 +1,64 @@
+/* global clearTimeout*/
 var Promise = require('es6-promise').Promise;
 var browserify = require('browserify');
 var path = require('path');
+var extend = require('util')._extend;
 var watchify = require('watchify');
 var fs = require('../utils/fs');
 var File = require('../utils/file');
 var log = require('../utils/log');
+
+function mapExternalFiles(vendorBundle) {
+    if (!vendorBundle) return;
+    return vendorBundle.map(function (v) {
+        return (typeof v === 'string') ? v : v.expose;
+    });
+}
+
+function bundle(b, fileObj, outFile) {
+    var b_ws = fs.createWriteStream(path.resolve(outFile));
+    b.bundle().pipe(b_ws);
+    return new Promise(function(resolve, reject) {
+        b_ws.end = function(){
+            //todo: verbose mode?
+            //log.info(' * ' + outFile + ' saved' );
+            fileObj.path = outFile;
+            return resolve(fileObj);
+        };
+        b_ws.on('error', reject);
+    });
+}
+
+function Watch (location, destination, options) {
+    this.options = options || {};
+    this.options.entries = options.fileObj.path;
+    this.fileObj = options.fileObj;
+    this.browserSync = options.browserSync;
+    this.outFile = path.join(destination, this.fileObj.relativeDir, this.fileObj.name);
+    var b = browserify(options, watchify.args);
+    var vendor = mapExternalFiles(this.options.vendorBundle);
+    if (vendor){
+        b.external(vendor);
+    }
+    b.require(this.fileObj.path, {expose: this.fileObj.name.split('.')[0]});
+        return this.file(b);
+}
+
+Watch.prototype.reloadOnce = function() {
+    clearTimeout(this.reloadTimout);
+    this.reloadTimout = setTimeout(this.browserSync.reload, 150);
+};
+
+Watch.prototype.file = function(b) {
+    var self = this;
+    this.w = watchify(b); //, { delay: 1000 }
+    log.info('   * watch ' + this.fileObj.relativeDir + this.fileObj.name ); //+ fileObj.name
+    this.w.on('update', function () {
+            bundle(self.w, self.fileObj, self.outFile).then(self.reloadOnce.bind(self)).catch(log.onError);
+    });
+    this.w.bundle();
+};
+
 
 function Browserify(location, destination, options){
     this.location = location;
@@ -39,66 +93,29 @@ Browserify.prototype.buildVendor = function(fileObj, options){
     });
 };
 
-Browserify.prototype.mapExternalFiles = function() {
-    if (!this.options.vendorBundle) return;
-    return this.options.vendorBundle.map(function (v) {
-        return (typeof v === 'string') ? v : v.expose;
-    });
-};
-
 Browserify.prototype.file = function(fileObj, browserSync) {
-    var self = this;
     var options = this.options || {};
     options.entries = fileObj.path;
     var b = browserify(options, (browserSync) ? watchify.args : undefined);
-    var vendor = this.mapExternalFiles();
+    var vendor = mapExternalFiles(this.options.vendorBundle);
     if (vendor){
         b.external(vendor);
     }
     b.require(fileObj.path, {expose: fileObj.name.split('.')[0]});
-    if (browserSync){
-        return self.watchBundle(b, fileObj, browserSync);
-    } else {
-        return self.bundle(b, fileObj);
-    }
-};
-
-Browserify.prototype.watchBundle = function(b, fileObj, browserSync) {
-    var self = this;
-    b = watchify(b, { delay: 1000 });
-    return new Promise(function(resolve, reject) {
-        b.on('update', function () {
-            self.bundle(b, fileObj).then(function(){
-                browserSync.reload();
-                resolve();
-            });
-        });
-        b.bundle();
-    });
-};
-
-Browserify.prototype.bundle = function(b, fileObj) {
-    var outFile = path.join(this.destination, fileObj.relativeDir, fileObj.name);
-    var b_ws = fs.createWriteStream(path.resolve(outFile));
-    b.bundle().pipe(b_ws);
-    return new Promise(function(resolve, reject) {
-        b_ws.end = function(){
-            //todo: verbose mode?
-            //log.info(' * ' + fileObj.name + ' saved in ' + self.destination);
-            fileObj.path = outFile;
-            return resolve(fileObj);
-        };
-        b_ws.on('error', reject);
-    });
+    return bundle(b, fileObj, path.join(this.destination, fileObj.relativeDir, fileObj.name));
 };
 
 Browserify.prototype.watch = function(browserSync) {
     var self = this;
     return fs.glob(this.location).then(function(fileObjs) {
         fileObjs.forEach(function (fileObj, i) {
-            self.file(fileObj, browserSync);
+            var options = extend(self.options || {},{
+                fileObj: fileObj,
+                browserSync:browserSync
+            });
+            new Watch(self.location, self.destination, options);
         });
-    });
+    }).catch(log.onError);
 };
 
 Browserify.prototype.write = function(){